feat: add new marketing components and enhance UI
- Introduced new marketing components: Hero, CTA, FAQ, FeatureGrid, Testimonials, and SiteHeader for improved user engagement.
- Added animated group functionality for smoother transitions.
- Updated button styles to include shadow effects for better visibility.
- Integrated Radix UI components for enhanced accessibility and usability.
- Included new images and optimized layout for a more appealing design.
