+++ package.json
@@ -23,6 +23,7 @@
   },
   "dependencies": {
     "@hookform/resolvers": "^5.2.1",
+    "@radix-ui/react-accordion": "^1.2.12",
     "@radix-ui/react-avatar": "^1.1.10",
     "@radix-ui/react-dialog": "^1.1.15",
     "@radix-ui/react-dropdown-menu": "^2.1.16",
@@ -36,6 +37,7 @@
     "@react-email/components": "0.1.1",
     "@react-email/html": "0.0.11",
     "@react-email/render": "^1.2.0",
+    "@studio-freight/react-lenis": "^0.0.47",
     "@tanstack/react-query": "^5.85.5",
     "@types/bcrypt": "^5.0.2",
     "@types/jsonwebtoken": "^9.0.10",
@@ -46,11 +48,13 @@
     "clsx": "^2.1.1",
     "cmdk": "^1.1.1",
     "date-fns": "4.1.0",
+    "framer-motion": "^12.23.16",
     "input-otp": "^1.4.2",
     "install": "^0.13.0",
     "jsonwebtoken": "^9.0.2",
     "lucide-react": "^0.525.0",
     "mongoose": "^8.17.2",
+    "motion": "^12.23.18",
     "next": "15.4.6",
     "next-themes": "^0.4.6",
     "nodemailer": "^7.0.5",
