--- src/app/(dashboard)/layout.tsx
@@ -1,11 +1,0 @@
-'use client';
-
-import ProtectedRoute from '@/components/ProtectedRoute';
-
-export default function DashboardLayout({
-  children,
-}: {
-  children: React.ReactNode;
-}) {
-  return <ProtectedRoute redirectTo="/log-in">{children}</ProtectedRoute>;
-}
