+++ src/app/(protected)/account/page.tsx
@@ -1,0 +1,5 @@
+import { AccountForm } from '@/app/(protected)/components/forms/account-form';
+
+export default function SettingsAccountPage() {
+  return <AccountForm />;
+}
