+++ src/app/(protected)/components/index.tsx
@@ -1,0 +1,4 @@
+export { AppSidebar } from './app-sidebar';
+export { DeleteUser } from './delete-user';
+export { UploadAvatar } from './upload-avatar';
+export { UserDropdown } from './user-dropdown';
