+++ src/app/(protected)/edit-profile/page.tsx
@@ -1,0 +1,5 @@
+import { ProfileForm } from '@/app/(protected)/components/forms/profile-form';
+
+export default function SettingsProfilePage() {
+  return <ProfileForm />;
+}
