+++ src/app/page.tsx
@@ -1,63 +1,33 @@
-import { Button } from '@/components/ui/button';
-import { Logo } from '@/components/ui/logo';
-import {
-  generateOrganizationStructuredData,
-  generateWebApplicationStructuredData,
-  generateWebsiteStructuredData,
-} from '@/seo/seo';
-import Link from 'next/link';
+import CTA from '@/components/marketing/CTA';
+import FAQ from '@/components/marketing/FAQ';
+import FeatureGrid from '@/components/marketing/FeatureGrid';
+import Footer from '@/components/marketing/Footer';
+import Hero from '@/components/marketing/Hero';
+import HowItWorks from '@/components/marketing/HowItWorks';
+import SiteHeader from '@/components/marketing/SiteHeader';
+import Testimonials from '@/components/marketing/Testimonials';
+import { siteConfig } from '@/seo/metadata';
+import type { Metadata } from 'next';
 
-const page = () => {
-  const websiteStructuredData = generateWebsiteStructuredData();
-  const organizationStructuredData = generateOrganizationStructuredData();
-  const webAppStructuredData = generateWebApplicationStructuredData();
-
-  return (
-    <>
-      {/* Structured Data */}
-      <script
-        type="application/ld+json"
-        dangerouslySetInnerHTML={{
-          __html: JSON.stringify(websiteStructuredData),
-        }}
-      />
-      <script
-        type="application/ld+json"
-        dangerouslySetInnerHTML={{
-          __html: JSON.stringify(organizationStructuredData),
-        }}
-      />
-      <script
-        type="application/ld+json"
-        dangerouslySetInnerHTML={{
-          __html: JSON.stringify(webAppStructuredData),
-        }}
-      />
-
-      <div className="flex flex-col items-center justify-center h-[100svh]">
-        <div className="text-center mb-8">
-          <Logo
-            width={80}
-            height={80}
-            variant="colored"
-            className="mx-auto mb-6"
-          />
-          <h1 className="text-4xl font-bold mb-2">Welcome to SaaS Starter</h1>
-          <p className="text-lg text-muted-foreground">
-            Modern SaaS foundation built with Next.js & TypeScript
-          </p>
-        </div>
-        <div className="flex gap-4">
-          <Button variant="outline" asChild>
-            <Link href="/log-in">Log in</Link>
-          </Button>
-          <Button variant="outline" asChild>
-            <Link href="/create-account">Create account</Link>
-          </Button>
-        </div>
-      </div>
-    </>
-  );
+export const metadata: Metadata = {
+  title: `Next Gen – ${siteConfig.name}`,
+  description: siteConfig.description,
+  alternates: { canonical: '/' },
 };
 
-export default page;
+export default function HomePage() {
+  return (
+    <>
+      <SiteHeader />
+      <main>
+        <Hero />
+        <FeatureGrid />
+        <HowItWorks />
+        <Testimonials />
+        <FAQ />
+        <CTA />
+      </main>
+      <Footer />
+    </>
+  );
+}
